requirements.txt
keras
tensorflow
numpy
opencv-python
scipy
imgaug
pillow
fonttools
tqdm